+++ bootstrap/lib/mix/tasks/nerves/env.ex
@@ -2,11 +2,22 @@ defmodule Mix.Tasks.Nerves.Env do
   use Mix.Task
   import Mix.Nerves.IO
 
-  @switches [info: :boolean]
+  @switches [info: :boolean, disable: :boolean]
 
   def run(argv) do
     debug_info "Env Start"
     {opts, _, _} = OptionParser.parse(argv, switches: @switches)
+    if opts[:disable] do
+      # This is the same code as Nerves.Env.disable/0, but I can't call that
+      # because we're about to load it immediately below. I also can't call it
+      # after it's loaded below because the act of compiling the nerves dep
+      # calls deps.precomile, which causes the system to get compiled unless
+      # Nerves.Env is disabled, which is what I'm trying to do here. :&
+      # This could probably be solved by moving Nerves.Env into nerves_bootstrap
+      System.put_env("NERVES_ENV_DISABLED", "1")
+    else
+      System.delete_env("NERVES_ENV_DISABLED")
+    end
     unless Code.ensure_compiled?(Nerves.Env) do
       Mix.Tasks.Deps.Compile.run ["nerves", "--include-children"]
     end