+++ lib/nerves/package/providers/docker.ex
@@ -105,8 +105,76 @@ defmodule Nerves.Package.Providers.Docker do
     |> File.rm_rf
   end
 
+  @doc """
+  Connect to a system configuration shell in a Docker container
+  """
+  @spec system_shell(Nerves.Package.t) :: :ok
+  def system_shell(pkg) do
+    container_name = preflight(pkg)
+    container_ensure_started(container_name)
+
+    # We need to get raw binary access to the stdout file descriptor
+    # so we can directly pass through control characters output by the command
+    stdout_port = Port.open({:fd, 0, 1}, [:binary, :eof, :stream, :out])
+
+    # We use the tty_sl driver for input because it handles tty geometry and
+    # streaming mode.
+    stdin_port = Port.open({:spawn, "tty_sl -c -e"}, [:binary, :eof, :stream, :in])
+
+    # We run the command through the script command to emulate a pty
+    cmd_port = Port.open({:spawn, "script -q /dev/null docker attach #{container_name}"}, [:binary, :eof, :stream, :stderr_to_stdout])
+
+    # Tell the script command about the terminal dimensions
+    {w, h} = get_tty_geometry(stdin_port)
+    Port.command(cmd_port, "stty rows #{h} cols #{w}\r")
+
+    platform_config = pkg.config[:platform_config][:defconfig]
+    defconfig = Path.join("/nerves/env/#{pkg.app}", platform_config)
+
+    create_build_cmd = [
+      "/nerves/env/platform/create-build.sh",
+      defconfig,
+      @working_dir,
+    ] |> Enum.join(" ")
+
+    Port.command(cmd_port, "#{create_build_cmd}\r")
+
+    system_shell_loop(stdin_port, stdout_port, cmd_port)
+  end
+
+  defp system_shell_loop(stdin_port, stdout_port, cmd_port) do
+    receive do
+      # Route input from stdin to the command port
+      {^stdin_port, {:data, data}} ->
+        Port.command(cmd_port, data)
+        system_shell_loop(stdin_port, stdout_port, cmd_port)
+
+      # Route output from the command port to stdout
+      {^cmd_port, {:data, data}} ->
+        Port.command(stdout_port, data)
+        system_shell_loop(stdin_port, stdout_port, cmd_port)
+
+      # If any of the ports get closed, break out of the loop
+      {_port, :eof} ->
+        :ok
+
+      # Ignore other messages
+      _message ->
+        system_shell_loop(stdin_port, stdout_port, cmd_port)
+    end
+  end
+
+  @ctrl_op_get_winsize 100
+
+  defp get_tty_geometry(tty_port) do
+    geometry = :erlang.port_control(tty_port, @ctrl_op_get_winsize, [])
+    |> :erlang.list_to_binary()
+    <<w::native-integer-size(32), h::native-integer-size(32)>> = geometry
+    {w, h}
+  end
+
   defp preflight(pkg) do
-    container_id() || create_container_id()
+    container_id(pkg) || create_container_id(pkg)
     name = container_name(pkg)
     _ = host_check()
     _ = config_check(pkg, name)
@@ -114,26 +182,29 @@ defmodule Nerves.Package.Providers.Docker do
   end
 
   defp container_name(pkg) do
-    if id = container_id() do
+    if id = container_id(pkg) do
       "#{pkg.app}-#{id}"
     end
   end
 
-  defp container_id() do
-    id_file = container_id_file()
+  defp container_id(pkg) do
+    id_file = container_id_file(pkg)
 
     if File.exists?(id_file) do
       File.read!(id_file)
+    else
+      create_container_id(pkg)
+      container_id(pkg)
     end
   end
 
-  defp container_id_file() do
-    Mix.Project.build_path
-    |> Path.join("nerves/.docker_id")
+  defp container_id_file(pkg) do
+    Artifact.base_dir(pkg)
+    |> Path.join(".docker_id")
   end
 
-  defp create_container_id() do
-    id_file = container_id_file()
+  defp create_container_id(pkg) do
+    id_file = container_id_file(pkg)
     id = Nerves.Utils.random_alpha_num(16)
     Path.dirname(id_file)
     |> File.mkdir_p!